Update client to support 64 bit card race
To match with https://github.com/edo9300/ygopro-core/pull/127

+++ gframe/data_manager.h
@@ -46,7 +46,7 @@ struct CardData {
 	uint32_t type;
 	uint32_t level;
 	uint32_t attribute;
-	uint32_t race;
+	uint64_t race;
 	int32_t attack;
 	int32_t defense;
 	uint32_t lscale;
@@ -60,7 +60,7 @@ struct CardDataC {
 	uint32_t type;
 	uint32_t level;
 	uint32_t attribute;
-	uint32_t race;
+	uint64_t race;
 	int32_t attack;
 	int32_t defense;
 	uint32_t lscale;
@@ -142,7 +142,7 @@ public:
 	std::wstring GetNumString(int num, bool bracket = false) const;
 	epro::wstringview FormatLocation(uint32_t location, int sequence) const;
 	std::wstring FormatAttribute(uint32_t attribute) const;
-	std::wstring FormatRace(uint32_t race, bool isSkill = false) const;
+	std::wstring FormatRace(uint64_t race, bool isSkill = false) const;
 	std::wstring FormatType(uint32_t type) const;
 	std::wstring FormatScope(uint32_t scope, bool hideOCGTCG = false) const;
 	std::wstring FormatSetName(const std::vector<uint16_t>& setcodes) const;
